requirements.txt
Flask
Flask-PyMongo
python-docx
pytesseract
Pillow
pymupdf
spacy
pandas
werkzeug
